refactor(api): split old grant process from flat
