+++ packages/api/src/adapters/inputs/http/association.http.ts
@@ -34,7 +34,6 @@ import { HttpErrorInterface, NotAssociationError } from "core";
 
 import associationService from "../../../modules/associations/associations.service";
 import grantService from "../../../modules/grant/grant.service";
-import { JoinedRawGrantDto } from "../../../modules/grant/@types/RawGrant";
 import associationIdentifierService from "../../../modules/association-identifier/association-identifier.service";
 import grantExtractService from "../../../modules/grant/grant-extract.service";
 import { errorHandler } from "../../../middlewares/error.middleware";
@@ -44,6 +43,7 @@ import applicationFlatService from "../../../modules/application-flat/applicatio
 import { PAYMENT_DTO_EXAMPLE, APPLICATION_DTO_EXAMPLE } from "./examples/Grants";
 import { OLD_APPLICATON_DTO, OLD_PAYMENT_DTO } from "./examples/OldGrants";
 import { DOCUMENT_DTO } from "./examples/Documents";
+import rawGrantService from "../../../modules/grant/raw-grant.service";
 
 export async function isAssoIdentifierFromAssoMiddleware(req, _res, next) {
     /*
@@ -262,7 +262,7 @@ export class AssociationHttp extends Controller {
         @Request() req,
     ): Promise<GetOldGrantsResponseDto> {
         const associationIdentifiers = req.assoIdentifier;
-        const grants = await grantService.getOldGrants(associationIdentifiers);
+        const grants = await rawGrantService.getOldGrants(associationIdentifiers);
         return { subventions: grants, count: grants.length };
     }
 
@@ -311,28 +311,6 @@ export class AssociationHttp extends Controller {
         stream.push(csv);
         stream.push(null);
         return stream;
-    }
-
-    /**
-     * Recherche les subventions liées à une association, format brut
-     *
-     * @deprecated test purposes
-     * @summary Recherche les subventions liées à une association, format brut
-     * @param identifier Identifiant Siren ou Rna
-     * @param req
-     */
-    @Deprecated()
-    @Get("/raw-grants")
-    @Security("jwt", ["admin"])
-    @Response<HttpErrorInterface>("404")
-    @Response<HttpErrorInterface>("403", "Accès refusé")
-    public async getRawGrants(
-        @Path() identifier: AssociationIdentifierDto,
-        @Request() req,
-    ): Promise<JoinedRawGrantDto[]> {
-        const associationIdentifiers = req.assoIdentifier;
-
-        return grantService.getRawGrantsDto(associationIdentifiers);
     }
 
     /**
